add : StatDB추가 및 initialize 유틸추가

+++ Assets/Util_Script/SaveData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [System.Serializable]
 public class SaveData 
 {
-    public SaveData(string name, int xp, int money, int hp, int happypoint)
+    public SaveData(string name, float xp, int money, float hp, float happypoint)
     {
         this.name = name;
         this.xp = xp;
@@ -16,10 +16,10 @@ public class SaveData
 
     // 임시로 public화 한거임 원래는 은닉성을 위해 private형이 맞음 
     public string name;
-    public int xp;
+    public float xp;
     public int money;
-    public int hp;
-    public int happypoint;
+    public float hp;
+    public float happypoint;
